add JKS configuration page
