+++ src-tauri/tauri.conf.json
@@ -16,7 +16,8 @@
         "all": true
       },
       "dialog": {
-        "save": true
+        "save": true,
+        "open": true
       }
     },
     "bundle": {
